Adds set_hash_list_size(...)
This internal function is used to set the size of the |hash_list|.
This comes in handy for check tests to test what happens when the
limit is hit.

Further, by setting the size of the |hash_list| in the tests, the
test-settings meson option is removed since there is no longer a need
to pass in MAX_GOP_LENGTH as define at compile time.

The test_checks.sh script and READMEs have been updated accordingly.