Adds a payload_buffer to store ongoing SEIs (#49)
* Adds a payload_buffer to store ongoing SEIs

When a SEI is about to be created everything up to the signature is
written. It is then completed when the signature exists. This means
that SEIs can pile up and storing pointers to these memory slots is
necessary for other signing plugins.

The members payload_ptr and signature_payload_ptr have been removed
since the are not used anymore.

Co-authored-by: bjornvolcker <bjornvolcker@users.noreply.github.com>