Adds a threaded plugin (#50)
The new plugin adds a dependency on threads and glib-2.0. It sets up
a worker thread from which the OpenSSL signing is called.

A new meson-option (signingplugin) to select between the two plugins is
added. There is an opation to use the threaded plugin unless we build with
libcheck (threaded_unless_check_dep). Note that the check tests only works
as intended with the unthreaded plugin.

Co-authored-by: bjornvolcker <bjornvolcker@users.noreply.github.com>