The public Axis root CA is not added to SEI
The certificate chain set by an Axis camera consists of three certificates.
The last one is the public root CA. This certificate is also added to
the library for ease of use when verifying the certificate chain.
Transmitting something that is already present in signed-video-framework
only increases the bitrate. Instead, this last certificate is added to the
certificate chain when decoding the tag. If the original chain did not use
the Axis root CA to sign verification of the chain will fail anyhow.

Tests have been updated.

Version bumped to v1.1.4