Timestamp information in SEI (#85)
New API to add a user specified timestamp with the nalu as an int64.
The timestamp is added to the general TLV, which is then signed. The
timestamp will therefore be signed and validated, but trusting this
timesstamp mean trusting in the clock and the user application adding
the correct timestamp.

New tests has been added to check the new API for validity and tests for
checking that the old API for adding nalu gives the same results as
adding a nalu with a NULL timestamp.