Timestamp information in SEI (#85) (#91)
New API to add a user specified timestamp (in int64_t) together with the
NALU. The timestamp is then added to the general TLV, which is then
signed. The timestamp will therefore be signed and validated, but trusting
this timestamp means trusting the clock and the user application adding
the correct timestamp.

The timestamp should be a Unix epoch in UTC format.

New tests has been added to check the new API for validity and tests for
checking that the old API for adding nalu gives the same results as
adding a nalu with a NULL timestamp.