Adds a queue of SEIs to sign
Puts hashes and signing requests in buffers. A hash to sign can only be
processed if the previous request is completed and has been fetched. If
not, SE is blocked and GOPs will be unsigned. By putting signing
requests and hashes in buffers, delays can be handled without loss of
SEIs.