Adds a nalu_str to latest_validation
this to be able to connect the validation status to the NALU type
of each validated item.

Adds a helper to get a character representation of most NALU types.
Further, refactors the API to get a validation string to also get
the nalu string controlled by a new enum.

Removed one unecessary memcpy in this process.

Bumps the version to v1.1.22.