Fixes potential segfault (#129)
when tearing down a signing thread. If signing is active when
closing down the plugin, memory might be freed that later is used.
This commit waits for an ongoing signing to complete before
freeing memory.

Further, now checks if there is available space in the buffer
before creating the SEI and sending for signing. Previously, the
check was done after.

Bumps to v1.1.28

---------

Co-authored-by: bjornvolcker <bjornvolcker@users.noreply.github.com>