Adds all intermediate certificates
For vendor Axis Communications, two intermediate certificates
are expected and a third root CA. Before this commit a check
for bitexactness of the root CA certificate was done, which
could cause missing out the certificate chain, even though
validation would pass.
Now the check only verifies there are in total three
certificates in the chain and the first two are added to the
SEI.

Bumps version to v1.1.29.