Adds a crypto TLV tag
A new TLV tag CRYPTO_INFO_TAG is created which is associated
with crypto information like the hashing algorithm used.
At this point the length of the data is assumed to be at most
255 bytes, hence one byte is enough to represent the length.

Both encode and decode functions are added and helper functions
are added to read and write the OID on serialized form, i.e.,
ASN.1/DER format.

A new struct to keep a message digest in both type and OID is
added. This will in particular help when a message digest for
the signature is added later.