Adds a setter for hash algorithm (#145)
* Adds a setter for hash algorithm

To customize usage of hash algorithms a new setter is added.
This new function sets the hashing algorithm, used to hash NAL
Units, given a name or OID that OpenSSL can interpret. If the
hash algorithm is not set, SHA256 is used.

A change of hash algorithm cannot be applied once the session has
started hashing NAL Units. Therefore a new member
|signing_started| has been added.

Basic signing tests is added.

NOTE that the validation side is not updated. The library still
assumes the video is hashed with SHA256. Transferring the
knowledge of which hash algorithm is in use is done in separate
commits.

Also, updates meson-build.yml

---------

Co-authored-by: bjornvolcker <bjornvolcker@users.noreply.github.com>