Added a new API to enable the golden SEI principle.
While using the principle, after the golden SEI is generated, the rest
of the SEIs only need to carry mandatory information.Modifications are added
to signed_video_h26x_auth.c and signed_video_h26x_sign.c for principle to work.
A new test is added to authentication side confirm the principle.