Adds AV1 support
Implementation is not fully commented and the validation side
tests have some uncertain parts, which needs to be revisited.
The validation side acts as expected for all corner cases, but
can be a little bit restricted and signal invalid for a longer
time.

The flag has_reference_hash set to true, which makes sure the first
I-frame is hashed both with buddy and copied to ref. For AV1 this
is crucial, since the ongoing hash update has to be finalized and
then copied to ref. This could affect H.26x.

For B-frames, AV1 adds an extra Frame Header (FH) is added to mark
where the existing P-frame is to be rendered. This FH is hashed
as a P-frame which will add an extra character in the NALU string
if looking at the validation results.