Fixes a bug in the threaded plugin running in central mode (#222)
If a session is closed while a hash is being signed the generated
signature returns after teardown. The signature is then still
added to the plugin output buffer which could be picked up by a
new session.

This commit solves that issue by only adding generated signatures
if the session is still active. Further, new session ids are now
unique until the number wraps around.

---------

Co-authored-by: bjornvolcker <bjornvolcker@users.noreply.github.com>