Fixes corner case with 2 unusable starting SEIs
Tho be able to support this a larger structural change is made.
Before the gop counter is in sync, each SEI is tested until it is
no longer feasible to test new GOPs.
Some flags and validation status are therefore stored as
intermediate temporary variables. A new function that updates the
"real" flags has been added.

One effect is that before the stream is synched with the SEIs,
twice as many "has signature" reports are output. This does not
affect the validation result.

A test for this new case has been added.