Remove unused second_hash and has_reference_hash members (#289)
Remove unused second_hash and has_reference_hash members

The second_hash member is no longer required as linking is now handled
with a separate hash in the TLV instead of being part of the hash list.
Consequently, the I-frame is now hashed in only one way. Additionally,
the has_reference_hash member has been removed as part of this cleanup.