Updates use_in_gop_hash wrt partial GOPs
When preparing for validation the number of BUs marked now
depends on the number of sent BUs if a partial GOP was triggered.
If validation fails, the number of marked items are extended to
avoid aborting validation too early.
Last, when verifying individual hashes the number of verified
ones will not exceed the number sent.