Renames internal files (#401)
When bringing in ONVIF Media Signing as a submodule
some functions share the same name. This will cause
multiple definitions if building the library with a
populated submodule, like FilePlayer will.

This commit prepends internal functions with "sv_"
to make names unique. This only applies to function
names common to ONVIF Media Signing.