Passes on threaded signing plugin if Media Signing is known (#412)
If ONVIF Media Signing is known (installed or present in the submodule)
the threaded signing plugin should pass on all operations to the
threaded plugin in media-signing-framework. This ensures that only one
thread for signing is setup, and that thread is "controlled" from
media-signing-framework.
NOTE: that using the plugin in media-signing-framework has nothing to
do whether the stream has been signed with ONVIF Media Signing format
or not.