Adds function to initialize ONVIF (#416)
If it is possible to sign using ONVIF Media Signing an instance is
created and all, already set, settings are ported. Media Signing is
can be used if the signing key has been provisioned in factory and
the codec is H.264 or H.265.
If the instantiation fails, the session falls back to Signed Video.