Adds a timestamp to the SEI TLV
Currently the SEI is associated with only one timestamp which
is the first Bitstream Unit of the (partial) GOP. In normal case
the I-frame. This means that, in neither the latest nor the
accumulated validation results, a correct time span for the
validated video can be displayed. By adding also an end timestamp
this can be achieved.