Correctly copies bu_info_t after processing at validation (#476)
If hashing of a Bitstream Unit is done on data without emulation
prevention bytes, that is, after them being removed, that data
is stored in a separate memory. This memory has to be copied as
well. This is critical for signing multiple GOPs to work on the
unsigned SEIs in the beginning of a session.