Increment of current_partial_gop done outside TLV (#487)
On the signing side, the current_partial_gop was updated
when encoding the TLVs. The TLV encoding process should
not take any actions. It should solely write the current
state into the SEI.

The update is now done outside encoding. Further, the
counter is initialized to -1 instead of 0. This makes
it more clear on the validation side when the first SEI
is to be used.

Co-authored-by: bjornvolcker <bjornvolcker@users.noreply.github.com>