Verifying with SEI returns OK also if SEI is not
The special case where a partial GOP is successfully validated,
but the SEI is not should return success. A failing signature
verification could be due to a bit flip of the SEI or a wrongly
set Public key. All the content could still be correct for
matching the GOP. This is in particular important when the SEI
is not in sync.